# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Monitor",
        "contact": {
            "name": "Glassen_IT"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://62.113.103.54/api",
            "description": "API Monitor"
        }
    ],
    "paths": {
...
        "/excludedpages/sites": {
            "get": {
                "tags": [
                    "excludedpages"
                ],
                "summary": "Список исключенных сайтов",
                "operationId": "9a669b0b49c18d2b3b791058c79f0ab5",
                "parameters": [
                    {
                        "name": "start(0)",
                        "in": "query",
                        "description": "Отсчет для пагинации",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int"
                        }
                    },
                    {
                        "name": "limit(20)",
                        "in": "query",
                        "description": "Лимит пагинации",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "url": {
                                                        "type": "string"
                                                    },
                                                    "page_count": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 5762,
                                                    "name": 1234,
                                                    "url": 1234,
                                                    "page_count": 1234
                                                },
                                                {
                                                    "id": "...",
                                                    "name": "...",
                                                    "url": "...",
                                                    "page_count": "..."
                                                }
                                            ]
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "count": {
                                                        "type": "integer"
                                                    },
                                                    "start": {
                                                        "type": "integer"
                                                    },
                                                    "limit": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": {
                                                "count": 10000,
                                                "start": 0,
                                                "limit": 20
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResult401"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/sites": {
            "get": {
                "tags": [
                    "sites"
                ],
                "summary": "Список сайтов",
                "operationId": "25a759f5d65767a3c78bcbc5a9de2c7f",
                "parameters": [
                    {
                        "name": "start(0)",
                        "in": "query",
                        "description": "Отсчет для пагинации",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int"
                        }
                    },
                    {
                        "name": "limit(20)",
                        "in": "query",
                        "description": "Лимит пагинации",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "url": {
                                                        "type": "string"
                                                    },
                                                    "page_count": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 5762,
                                                    "name": "ExampleSite",
                                                    "url": "examplesite.com",
                                                    "page_count": 99
                                                },
                                                {
                                                    "id": "...",
                                                    "name": "...",
                                                    "url": "...",
                                                    "page_count": "..."
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResult401"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "sites"
                ],
                "summary": "Добавлените сайтов",
                "operationId": "73917f8e0c43d619d5a89ccb29c37044",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "urls": {
                                        "example": [
                                            "site1.com",
                                            "site2.com",
                                            "..."
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "example": {
                                                "message": "Запрос на добавление и парсинг сайтов создан"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResult401"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/sites/scan": {
            "post": {
                "tags": [
                    "sites"
                ],
                "summary": "Полное сканирование сраниц(ы)",
                "operationId": "eec104f4b928b077eadd1f70356d68cb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": "123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "example": {
                                                "message": "Запрос на парсинг сайтов создан"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResult401"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sites/search": {
            "post": {
                "tags": [
                    "sites"
                ],
                "summary": "Поиск новых страниц",
                "operationId": "bf9c66a4b01807cd56e5c8296cc7a520",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": "123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "example": {
                                                "message": "Запрос на поиск новых страниц создан"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResult401"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ErrorResult401": {
                "title": "Шаблон ошибки статуса авторизации",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "message": "Unauthenticated."
                }
            }
        }
    },
...
}